+++ src/test/java/edu/illinois/library/cantaloupe/source/OkHttpHTTPImageInputStreamClientTest.java
@@ -1,11 +1,9 @@
 package edu.illinois.library.cantaloupe.source;
 
-import edu.illinois.library.cantaloupe.http.Headers;
 import edu.illinois.library.cantaloupe.http.Range;
 import edu.illinois.library.cantaloupe.http.Response;
 import edu.illinois.library.cantaloupe.test.BaseTest;
 import edu.illinois.library.cantaloupe.test.WebServer;
-import okhttp3.OkHttpClient;
 import org.eclipse.jetty.server.Request;
 import org.eclipse.jetty.server.handler.DefaultHandler;
 import org.junit.jupiter.api.AfterEach;
@@ -21,12 +19,14 @@ public class OkHttpHTTPImageInputStreamClientTest extends BaseTest {
 
     private WebServer server;
 
+    @Override
     @BeforeEach
     public void setUp() throws Exception {
         super.setUp();
         server = new WebServer();
     }
 
+    @Override
     @AfterEach
     public void tearDown() throws Exception {
         super.tearDown();
@@ -34,7 +34,7 @@ public class OkHttpHTTPImageInputStreamClientTest extends BaseTest {
     }
 
     @Test
-    void testSendHEADRequest() throws Exception {
+    void sendHEADRequest() throws Exception {
         server.setHandler(new DefaultHandler() {
             @Override
             public void handle(String target,
@@ -48,17 +48,39 @@ public class OkHttpHTTPImageInputStreamClientTest extends BaseTest {
         });
         server.start();
 
-        final String uri = server.getHTTPURI().toString();
-        final OkHttpClient httpClient = new OkHttpClient();
-
+        final HTTPRequestInfo requestInfo =
+                new HTTPRequestInfo(server.getHTTPURI().toString());
         final OkHttpHTTPImageInputStreamClient instance =
-                new OkHttpHTTPImageInputStreamClient(httpClient, uri);
+                new OkHttpHTTPImageInputStreamClient(requestInfo);
 
         instance.sendHEADRequest();
     }
 
     @Test
-    void testSendHEADRequestSendsExtraHeaders() throws Exception {
+    void sendHEADRequestSendsRequestInfoCredentials() throws Exception {
+        server.setHandler(new DefaultHandler() {
+            @Override
+            public void handle(String target,
+                               Request baseRequest,
+                               HttpServletRequest request,
+                               HttpServletResponse response) {
+                assertEquals("Basic dXNlcjpzZWNyZXQ=",
+                        baseRequest.getHeader("Authorization"));
+                baseRequest.setHandled(true);
+            }
+        });
+        server.start();
+
+        final HTTPRequestInfo requestInfo = new HTTPRequestInfo(
+                server.getHTTPURI().toString(), "user", "secret");
+        final OkHttpHTTPImageInputStreamClient instance =
+                new OkHttpHTTPImageInputStreamClient(requestInfo);
+
+        instance.sendHEADRequest();
+    }
+
+    @Test
+    void sendHEADRequestSendsRequestInfoHeaders() throws Exception {
         server.setHandler(new DefaultHandler() {
             @Override
             public void handle(String target,
@@ -71,27 +93,24 @@ public class OkHttpHTTPImageInputStreamClientTest extends BaseTest {
         });
         server.start();
 
-        final String uri = server.getHTTPURI().toString();
-        final OkHttpClient httpClient = new OkHttpClient();
+        final HTTPRequestInfo requestInfo =
+                new HTTPRequestInfo(server.getHTTPURI().toString());
+        requestInfo.getHeaders().add("X-Cats", "yes");
 
         final OkHttpHTTPImageInputStreamClient instance =
-                new OkHttpHTTPImageInputStreamClient(httpClient, uri);
-        Headers extraHeaders = new Headers();
-        extraHeaders.add("X-Cats", "yes");
-        instance.setExtraRequestHeaders(extraHeaders);
+                new OkHttpHTTPImageInputStreamClient(requestInfo);
 
         instance.sendHEADRequest();
     }
 
     @Test
-    void testSendGETRequest() throws Exception {
+    void sendGETRequest() throws Exception {
         server.start();
 
-        final String uri = server.getHTTPURI().toString() + "/jpg";
-        final OkHttpClient httpClient = new OkHttpClient();
-
+        final HTTPRequestInfo requestInfo =
+                new HTTPRequestInfo(server.getHTTPURI() + "/jpg");
         final OkHttpHTTPImageInputStreamClient instance =
-                new OkHttpHTTPImageInputStreamClient(httpClient, uri);
+                new OkHttpHTTPImageInputStreamClient(requestInfo);
 
         Response response = instance.sendGETRequest(new Range(0, 1, 4));
 
@@ -101,7 +120,30 @@ public class OkHttpHTTPImageInputStreamClientTest extends BaseTest {
     }
 
     @Test
-    void testSendGETRequestSendsExtraHeaders() throws Exception {
+    void sendGETRequestSendsRequestInfoCredentials() throws Exception {
+        server.setHandler(new DefaultHandler() {
+            @Override
+            public void handle(String target,
+                               Request baseRequest,
+                               HttpServletRequest request,
+                               HttpServletResponse response) {
+                assertEquals("Basic dXNlcjpzZWNyZXQ=",
+                        baseRequest.getHeader("Authorization"));
+                baseRequest.setHandled(true);
+            }
+        });
+        server.start();
+
+        final HTTPRequestInfo requestInfo = new HTTPRequestInfo(
+                server.getHTTPURI() + "/jpg", "user", "secret");
+        final OkHttpHTTPImageInputStreamClient instance =
+                new OkHttpHTTPImageInputStreamClient(requestInfo);
+
+        instance.sendGETRequest(new Range(0, 1, 4));
+    }
+
+    @Test
+    void sendGETRequestSendsRequestInfoHeaders() throws Exception {
         server.setHandler(new DefaultHandler() {
             @Override
             public void handle(String target,
@@ -114,14 +156,11 @@ public class OkHttpHTTPImageInputStreamClientTest extends BaseTest {
         });
         server.start();
 
-        final String uri = server.getHTTPURI().toString() + "/jpg";
-        final OkHttpClient httpClient = new OkHttpClient();
-
+        final HTTPRequestInfo requestInfo =
+                new HTTPRequestInfo(server.getHTTPURI() + "/jpg");
+        requestInfo.getHeaders().add("X-Cats", "yes");
         final OkHttpHTTPImageInputStreamClient instance =
-                new OkHttpHTTPImageInputStreamClient(httpClient, uri);
-        Headers extraHeaders = new Headers();
-        extraHeaders.add("X-Cats", "yes");
-        instance.setExtraRequestHeaders(extraHeaders);
+                new OkHttpHTTPImageInputStreamClient(requestInfo);
 
         instance.sendGETRequest(new Range(0, 1, 4));
     }